updated project structure added folders renamed notification-center to notifications
