+++ gulpfile.js
@@ -90,24 +90,24 @@ const refactoredSrc = [
   "./src/js/cloud-functions.js",
   "./src/js/misc.js",
   "./src/js/layouts.js",
-  "./src/js/monkey.js",
+  "./src/js/elements/monkey.js",
   "./src/js/result-filters.js",
-  "./src/js/notification-center.js",
-  "./src/js/leaderboards.js",
+  "./src/js/elements/notifications.js",
+  "./src/js/elements/leaderboards.js",
   "./src/js/sound.js",
-  "./src/js/custom-text.js",
-  "./src/js/shift-tracker.js",
+  "./src/js/test/custom-text.js",
+  "./src/js/test/shift-tracker.js",
   "./src/js/test/test-stats.js",
   "./src/js/theme-colors.js",
   "./src/js/test/out-of-focus.js",
   "./src/js/chart-controller.js",
   "./src/js/theme-controller.js",
   "./src/js/test/caret.js",
-  "./src/js/custom-text-popup.js",
-  "./src/js/manual-restart-tracker.js",
+  "./src/js/popups/custom-text-popup.js",
+  "./src/js/test/manual-restart-tracker.js",
   "./src/js/config.js",
   "./src/js/test/focus.js",
-  "./src/js/account-icon.js",
+  "./src/js/elements/account-icon.js",
   "./src/js/test/practise-missed.js",
   "./src/js/test/test-ui.js",
   "./src/js/test/keymap.js",
@@ -119,7 +119,7 @@ const refactoredSrc = [
   "./src/js/test/test-logic.js",
   "./src/js/test/funbox.js",
   "./src/js/test/pace-caret.js",
-  "./src/js/quote-search-popup.js",
+  "./src/js/popups/quote-search-popup.js",
   "./src/js/tag-controller.js",
   "./src/js/ui.js",
   "./src/js/test/pb-crown.js",
@@ -129,15 +129,15 @@ const refactoredSrc = [
   "./src/js/commandline-lists.js",
   "./src/js/commandline.js",
   "./src/js/challenge-controller.js",
-  "./src/js/custom-word-amount-popup.js",
-  "./src/js/custom-test-duration-popup.js",
+  "./src/js/popups/custom-word-amount-popup.js",
+  "./src/js/popups/custom-test-duration-popup.js",
   "./src/js/test/test-config.js",
-  "./src/js/loader.js",
+  "./src/js/elements/loader.js",
   "./src/js/mini-result-chart.js",
-  "./src/js/sign-out-button.js",
-  "./src/js/layout-emulator.js",
-  "./src/js/version-popup.js",
-  "./src/js/support-popup.js",
+  "./src/js/elements/sign-out-button.js",
+  "./src/js/test/layout-emulator.js",
+  "./src/js/popups/version-popup.js",
+  "./src/js/popups/support-popup.js",
 ];
 
 //legacy files
