+++ src/js/elements/account-icon.js
