+++ src/js/elements/sign-out-button.js
