+++ src/js/global-dependencies.js
@@ -12,7 +12,7 @@ import * as Misc from "./misc";
 import * as CloudFunctions from "./cloud-functions";
 import layouts from "./layouts";
 import * as Monkey from "./monkey";
-import * as Notifications from "./notification-center";
+import * as Notifications from "./notifications";
 import * as ResultFilters from "./result-filters";
 import * as Leaderboards from "./leaderboards";
 import * as Sound from "./sound";
